requirements.txt
fabric==2.6.0
python-dotenv==0.19.1